pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "body-align"
version = "0.2.0"
description = "CLI tool to align progress photos using MediaPipe pose detection and rembg background removal"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["fitness", "progress", "photos", "alignment", "mediapipe", "pose"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
]
dependencies = [
    "opencv-python>=4.8",
    "mediapipe>=0.10",
    "numpy>=1.24",
    "rembg[cpu]>=2.0",
    "pillow>=10.0",
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
body-align = "body_align.cli:main"

[project.urls]
Homepage = "https://github.com/jStrider/body-progress-tracker"
Repository = "https://github.com/jStrider/body-progress-tracker"
Issues = "https://github.com/jStrider/body-progress-tracker/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["body_align*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
